requirements.txt
# Python version
python==3.10.0

# Dependencies
farm-haystack[colab,inference]
requests
ipython
opencv-python